refactor(metrics): consolidate module structure and rename cleanup tracker
- collapse export, flow, processing, userspace, and k8s label enum files into labels.rs
- absorb error.rs into server.rs (single consumer)
- rename MetricCleanupTracker → MetricsEvictor, move to evictor.rs
- remove debug_enabled from MetricsEvictor (policy now owned by main)
- unify evictor creation and task registration into a single block in main
- remove debug/trace log artifacts, narrating comments, and dead code
- drop invalid unit tests that relied on shared OnceLock global state
