+++ mermin/src/metrics.rs
@@ -1,11 +1,6 @@
-pub mod cleanup;
 pub mod ebpf;
-pub mod error;
-pub mod export;
-pub mod flow;
-pub mod k8s;
+pub mod evictor;
+pub mod labels;
 pub mod opts;
-pub mod processing;
 pub mod registry;
 pub mod server;
-pub mod userspace;
