--- mermin/src/metrics/export.rs
@@ -1,37 +1,0 @@
-//! Enums for export-related metrics labels.
-
-/// Export status for flow spans.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum ExportStatus {
-    Ok,
-    Error,
-    NoOp,
-}
-
-impl ExportStatus {
-    pub const fn as_str(self) -> &'static str {
-        match self {
-            ExportStatus::Ok => "ok",
-            ExportStatus::Error => "error",
-            ExportStatus::NoOp => "noop",
-        }
-    }
-}
-
-/// Exporter name for metrics.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum ExporterName {
-    Otlp,
-    Stdout,
-    Noop,
-}
-
-impl ExporterName {
-    pub const fn as_str(self) -> &'static str {
-        match self {
-            ExporterName::Otlp => "otlp",
-            ExporterName::Stdout => "stdout",
-            ExporterName::Noop => "noop",
-        }
-    }
-}
